requirements.txt
# Core dependencies
requests>=2.25.0
python-dotenv>=0.19.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.10.0

# Package building and publishing
build>=0.7.0
twine>=4.0.0
setuptools>=61.0
wheel>=0.37.0
